Got most notifications implemented
I have the open, modified, and closed notifications implemented;
these all send the appropriate jsonrpc notifications to the agent.
I don't have any tests for them yet, but I can see that it's getting
the right outputs in the *cody events* buffer with the agent's stderr.

I've also made a pretty serious attempt at implementing following
the cursor and the selection (i.e. point and optional mark) as they
change across buffers we're "tracking" with Cody. However, it's not
quite working yet. I'm losing the buffer-local values of the hook
callback functions. But I wanted to check in progress so far.