made completions object-oriented
Our data structures for representing completions in elisp
were becoming ad-hoc and difficult to understand and modify.

The completions UX is going to need many new features, so to
keep things more readable I introduced three new classes.

The old code was just bolting a bunch of computed metadata
onto the jsonrpc response object from the Agent, and now it
all goes into named slots with readable accessors, for the
most part.

I also fixed a bug where M-n and M-p were trying to cycle
the completion everywhere in the buffer. Changed it so that
it only cycles completions when you are at the beginning of
an active, acceptable completion.

Telemetry
---------

The telemetry wasn't working before this change, and it's
still not working now. However, teej informs me that it
will largely move into the Agent next week, so I'm going
to be removing all that code next week in any event.

Testing
-------
We still don't have tests yet. I want to start writing them,
perhaps with ERT (built into Emacs). But I want to get this
architectural change out quickly so I unblock anyone who
wants to do nontrivial work on the file concurrently.