Added automatic completion triggering for Emacs
This commit introduces automatic completion triggering for Cody-Emacs.
It's experimental at best, and is quite over-zealous. But it works.

I'm mostly committing now because it's a milestone.  I'd say it's not
really usable yet. Mainly since I still need to implement the "patching"
logic that Valery explained to me, where it needs to replace parts of
the line prefix and suffix. What I have now is a hack that only works
in a few syntactic positions. But also it needs a lot of fine-tuning.

Other stuff in this commit:

  * Got custom menu support working in the Cody mode-line icon
    - not super satisfied with the solution, but it works
      - (I spent way longer than I want to admit on this)
    - the menu itself is mostly a placeholder, but easy to add to

  * misc code cleanup & minor bug fixes