Fixed agent not starting up
Something caused an empty-string argument, which we we were passing
to node when starting the agent, to have the agent error out with
"unknown command: ''".

The empty string command was added back in Montreal and worked fine
for quite some time, so I'm not sure what caused it to break.