fixed a background error
The byte compiler spotted a bug where I was trying to
call a function that I had deleted, when an automatic
completion suggestion contained zero results. It was
not messaging anything to the user, but may have been
logging to a jsonrpc log somewhere.

I removed the function for now; we can figure out the
right UX for this case later.