reverted diff cache test function
Turns out switching the test to 'eq broke completion display
in practice -- it was claiming that the diffs included some
deletions and/or overwrites, when it should be insertions only.
Putting the test back to 'equal fixes it.

I'm guessing it's because of this:

       ;; rest
       (a `((- . ,a)))
       (b `((+ . ,b)))))))

which is doing consing, so we need to check for structural equality.