update agent dist
The latest version of agent doesn't depend on keytar, which means we can
remove the darwin only keytar file. This allows me to use cody on linux.

I created this from commit 266cef342 (2024-07-05)

Note: there are a bunch of files I see in the agent/dist dir (mostly
wasm) that I assume we also need? For now I am matching what was in this
dir before.